Modularized Scrolling
This is a pretty crude, but functional implementation of the
modularized scrolling from
http://www.onextrapixel.com/2013/09/18/onepagescroll-js-creating-an-appl
es-iphone-5s-website/.

It would be good to organize things a bit more (you know, make separate
css and js folders, etc.) in the near future.

We can fill in the material over the sketches I made as we go along.